refactor(rag-knowledge): U1 部分 — dto と smart_chunk を新パッケージに移設 (#730)
Issue #730 (RAGKnowledgeService / PipelineController の Port 分離) の Unit 1
の部分実装。dto と汎用チャンキング関数を新しい配置に移し、`rag_knowledge.py`
を新配置から re-export する形に縮小する。

- 新規: `src/rag/search/{__init__.py, models.py}`
  - dto: RAGRetrievalResult / VectorSearchItem / BM25SearchItem / RawSearchResults
- 新規: `src/rag/admin/{__init__.py, models.py}`
  - dto: DocumentResult
- 新規: `src/rag/indexer/smart_chunking.py`
  - 関数: smart_chunk
- 改訂: `src/rag/rag_knowledge.py`
  - dto / smart_chunk の元定義を削除し、新配置から import
  - 1078 LOC → 約 880 LOC
- 改訂: `src/rag/cli.py:928`
  - smart_chunk の import 元を新配置に更新

format 関数（format_raw_search_results / format_document_response /
format_file_size）と `RAGKnowledgeService` クラス本体は U1 残作業 / U2 で
対応する。
