+++ src/rag/admin/__init__.py
@@ -1,0 +1,5 @@
+"""管理系 Port + Adapter（Source / Stats / Document）.
+
+仕様: docs/specs/rag-knowledge.md / docs/specs/search-response.md /
+      docs/specs/infrastructure/content-listing.md
+"""
