+++ src/rag/rag_knowledge.py
@@ -10,16 +10,19 @@ import hashlib
 import logging
 import mimetypes
 from collections.abc import Sequence
-from dataclasses import dataclass, field
 from pathlib import Path, PurePosixPath
 from typing import TYPE_CHECKING, Any, cast
 from urllib.parse import urldefrag
 
-from .chunker import chunk_text
-from .content_detector import ContentType, detect_content_type
+from .admin.models import DocumentResult
 from .converter.converter import get_converted_rel_path
-from .heading_chunker import chunk_by_headings
-from .table_chunker import chunk_table_data
+from .indexer.smart_chunking import smart_chunk
+from .search.models import (
+    BM25SearchItem,
+    RAGRetrievalResult,
+    RawSearchResults,
+    VectorSearchItem,
+)
 from .vector_store import DocumentChunk, VectorStore
 
 if TYPE_CHECKING:
@@ -27,179 +30,6 @@ if TYPE_CHECKING:
     from .hybrid_search import HybridSearchEngine
 
 logger = logging.getLogger(__name__)
-
-
-def smart_chunk(
-    text: str, chunk_size: int, chunk_overlap: int,
-) -> list[tuple[str, str]]:
-    """コンテンツタイプに応じた適切なチャンキング手法を選択する.
-
-    仕様: docs/specs/rag-knowledge.md
-
-    - TABLE: テーブルデータとして行単位でチャンキング
-    - HEADING/MIXED: 見出し単位でチャンキング
-    - PROSE: 従来の段落ベースチャンキング
-
-    Args:
-        text: チャンキング対象のテキスト
-        chunk_size: チャンクの最大文字数
-        chunk_overlap: チャンク間のオーバーラップ文字数
-
-    Returns:
-        (content, section_path) のタプルリスト
-    """
-    if not text or not text.strip():
-        return []
-
-    content_type = detect_content_type(text)
-    logger.debug("Detected content type: %s", content_type.value)
-
-    if content_type == ContentType.TABLE:
-        table_chunks = chunk_table_data(text, row_context_size=1, max_chunk_size=chunk_size)
-        if table_chunks:
-            return [(c.content, c.section_path) for c in table_chunks]
-        logger.debug("Table chunking returned no results, falling back to prose")
-
-    if content_type in (ContentType.HEADING, ContentType.MIXED):
-        heading_chunks = chunk_by_headings(text, max_chunk_size=chunk_size, min_chunk_size=max(1, chunk_size // 4))
-        if heading_chunks:
-            return [(c.content, c.section_path) for c in heading_chunks]
-        logger.debug("Heading chunking returned no results, falling back to prose")
-
-    return [
-        (c, "") for c in chunk_text(
-            text, chunk_size=chunk_size, chunk_overlap=chunk_overlap,
-        )
-    ]
-
-
-@dataclass
-class RAGRetrievalResult:
-    """RAG検索結果.
-
-    仕様: docs/specs/rag-knowledge.md
-
-    Attributes:
-        context: フォーマット済みテキスト（システムプロンプト注入用）
-        sources: ユニークなソースURLリスト（表示用）
-    """
-
-    context: str
-    sources: list[str]
-
-
-@dataclass
-class VectorSearchItem:
-    """ベクトル検索の生結果アイテム.
-
-    仕様: docs/specs/search-response.md
-
-    Attributes:
-        text: チャンクテキスト
-        source_url: ソースURL（source_id）
-        distance: cosine距離（0に近いほど類似）
-        chunk_index: チャンクインデックス（0始まり）
-        title: コンテンツのタイトル
-        source_type: ソース種別
-        total_chunks: 当該ソースのチャンク総数（0はレガシーデータ）
-        section_path: 見出し階層（> 区切り）
-    """
-
-    text: str
-    source_url: str
-    distance: float
-    chunk_index: int
-    title: str = ""
-    source_type: str = ""
-    total_chunks: int = 0
-    collected_at: str = ""
-    section_path: str = ""
-    url: str = ""
-
-    def to_dict(self) -> dict[str, object]:
-        """JSON 出力用 dict に変換する."""
-        d: dict[str, object] = {
-            "text": self.text,
-            "source_url": self.source_url,
-            "distance": self.distance,
-            "chunk_index": self.chunk_index,
-            "title": self.title,
-            "source_type": self.source_type,
-            "total_chunks": self.total_chunks,
-            "collected_at": self.collected_at,
-            "section_path": self.section_path,
-        }
-        if self.url:
-            d["url"] = self.url
-        return d
-
-
-@dataclass
-class BM25SearchItem:
-    """BM25検索の生結果アイテム.
-
-    仕様: docs/specs/search-response.md
-
-    Attributes:
-        text: チャンクテキスト
-        source_url: ソースURL（source_id）
-        score: BM25スコア（高いほどキーワード一致）
-        doc_id: ドキュメントID
-        chunk_index: チャンクインデックス（0始まり）
-        title: コンテンツのタイトル
-        source_type: ソース種別
-        total_chunks: 当該ソースのチャンク総数（0はレガシーデータ）
-        section_path: 見出し階層（> 区切り）
-        url: 元 URL（取得元の外部 URL。存在する場合のみ）
-    """
-
-    text: str
-    source_url: str
-    score: float
-    doc_id: str
-    chunk_index: int = 0
-    title: str = ""
-    source_type: str = ""
-    total_chunks: int = 0
-    collected_at: str = ""
-    section_path: str = ""
-    url: str = ""
-
-    def to_dict(self) -> dict[str, object]:
-        """JSON 出力用 dict に変換する."""
-        d: dict[str, object] = {
-            "text": self.text,
-            "source_url": self.source_url,
-            "score": self.score,
-            "doc_id": self.doc_id,
-            "chunk_index": self.chunk_index,
-            "title": self.title,
-            "source_type": self.source_type,
-            "total_chunks": self.total_chunks,
-            "collected_at": self.collected_at,
-            "section_path": self.section_path,
-        }
-        if self.url:
-            d["url"] = self.url
-        return d
-
-
-@dataclass
-class RawSearchResults:
-    """ベクトル検索・BM25検索の生結果（準Agentic Search用）.
-
-    仕様: docs/specs/rag-knowledge.md
-
-    統合パイプライン（min-max正規化 + CC結合）を迂回し、
-    各エンジンの生結果を個別にLLMに渡して判断を委譲する。
-
-    Attributes:
-        vector_results: ベクトル検索の生結果リスト
-        bm25_results: BM25検索の生結果リスト
-    """
-
-    vector_results: list[VectorSearchItem]
-    bm25_results: list[BM25SearchItem]
 
 
 def format_raw_search_results(raw: RawSearchResults) -> str:
@@ -790,33 +620,6 @@ class RAGKnowledgeService:
 _TEXT_EXTENSIONS: frozenset[str] = frozenset(
     {".md", ".txt", ".adoc", ".html", ".htm", ".json"}
 )
-
-
-@dataclass
-class DocumentResult:
-    """ドキュメント全文取得の結果.
-
-    仕様: docs/specs/search-response.md
-
-    Attributes:
-        source_id: ソース識別子
-        title: コンテンツのタイトル
-        source_type: ソース種別
-        format: 取得形式（"text" または "original"）
-        content: ドキュメントテキスト（バイナリの場合は情報文字列）
-        is_binary: バイナリファイルか否か
-        error: エラーメッセージ（エラー時のみ）
-    """
-
-    source_id: str
-    title: str
-    source_type: str
-    format: str
-    content: str
-    is_binary: bool = False
-    error: str | None = None
-    collected_at: str = ""
-    extra: dict[str, object] = field(default_factory=dict)
 
 
 _VALID_FORMATS: frozenset[str] = frozenset({"text", "original"})
