+++ src/rag/search/__init__.py
@@ -1,0 +1,4 @@
+"""検索系 Port + Adapter + 内部 dto.
+
+仕様: docs/specs/rag-knowledge.md / docs/specs/search-response.md
+"""
